requirements_advanced.txt
opencv-python==4.8.1.78
torch==2.0.1
torchvision==0.15.2
tensorflow==2.13.0
scikit-learn==1.3.0
Pillow==10.0.0
numpy==1.24.3
web3==6.9.0
eth-account==0.9.0
paho-mqtt==1.6.1
qrcode==7.4.2
scikit-image==0.21.0
pytesseract==0.3.10